Added additions to README.md and use of main module

+++ src/routes/validation/__init__.py
@@ -49,7 +49,16 @@ validation_blueprint: Blueprint = Blueprint('validation', __name__)
 logger: logging.Logger = logging.getLogger(__name__)
 
 
+def get_groups_token() -> str:
+    return request.headers.get('authorization')[7:]
+
+
 def check_metadata_upload():
+    """
+    Checks the uploaded file
+
+    Returns dict containing upload details or an 'error' key if something went wrong
+    """
     result: dict = {
         'error': None
     }
@@ -66,7 +75,16 @@ def check_metadata_upload():
     return result
 
 
-def set_file_details(pathname):
+def set_file_details(pathname: str):
+    """
+    Creates a dictionary of file and path details
+
+    Parameters
+    ----------
+    pathname str pathname
+
+    Returns dict containing the filename and fullpath details
+    """
     base_path = get_base_path()
     return {
         'pathname': pathname,
@@ -74,12 +92,31 @@ def set_file_details(pathname):
     }
 
 
-def get_metadata(path):
+def get_metadata(path: str) -> list:
+    """
+    Parses a tsv and returns the rows of that tsv
+
+    Parameters
+    ----------
+    path str path where the tsv file is stored
+
+    Returns list of dictionaries
+    """
     result = get_csv_records(path)
     return result.get('records')
 
 
-def validate_tsv(schema='metadata', path=None):
+def validate_tsv(schema='metadata', path=None) -> dict:
+    """
+    Calls methods of the Ingest Validation Tools submodule
+
+    Parameters
+    ----------
+    schema str name of the schema to validate against
+    path str path of the tsv for Ingest Validation Tools
+
+    Returns dict containing validation results
+    """
     try:
         schema_name = (
             schema if schema != 'metadata'
@@ -89,15 +126,32 @@ def validate_tsv(schema='metadata', path=None):
         result = {'Preflight': str(e)}
     else:
         try:
+            entity_url: str = commons_file_helper.ensureTrailingSlashURL(current_app.config['ENTITY_WEBSERVICE_URL'])
+            app_context: dict = {
+                'request_header': {'X-SenNet-Application': 'ingest-api'},
+                'entities_url': f"{entity_url}entities/"
+            }
             report_type = ingest_validation_tools_table_validator.ReportType.JSON
+            cedar_api_key: str = current_app.config['CEDAR_API_KEY']
             result = ingest_validation_tools_validation_utils\
-                .get_tsv_errors(path, schema_name=schema_name, report_type=report_type)
+                .get_tsv_errors(path, schema_name=schema_name, report_type=report_type, cedar_api_key=cedar_api_key,
+                                globus_token=get_groups_token(), app_context=app_context)
         except Exception as e:
             result = rest_server_err(e, True)
     return json.dumps(result)
 
 
-def create_tsv_from_path(path, row):
+def create_tsv_from_path(path: str, row: int) -> dict:
+    """
+    Creates a tsv from path of a specific row. This is in order to validate only one if necessary.
+
+    Parameters
+    ----------
+    path str ath of original tsv
+    row int row number in tsv to extract for new tsv
+
+    Returns dict containing file details
+    """
     try:
         records = get_csv_records(path, records_as_arr=True)
         result = set_file_details(f"{time.time()}.tsv")
@@ -112,7 +166,24 @@ def create_tsv_from_path(path, row):
     return result
 
 
-def determine_schema(entity_type, sub_type):
+def get_cedar_schema_ids() -> dict:
+    return {
+        'Block': '3e98cee6-d3fb-467b-8d4e-9ba7ee49eeff',
+        'Section': '01e9bc58-bdf2-49f4-9cf9-dd34f3cc62d7',
+        'Suspension': 'ea4fb93c-508e-4ec4-8a4b-89492ba68088'
+    }
+
+
+def check_cedar(entity_type: str, sub_type, upload) -> bool:
+    records = get_metadata(upload.get('fullpath'))
+    if len(records) > 0:
+        if equals(entity_type, "Sample") and 'metadata_schema_id' in records[0]:
+            cedar_sample_sub_type_ids = get_cedar_schema_ids()
+            return equals(records[0]['metadata_schema_id'], cedar_sample_sub_type_ids[sub_type])
+    return True
+
+
+def determine_schema(entity_type: str, sub_type) -> str:
     if equals(entity_type, "Sample"):
         if not sub_type:
             return rest_bad_req("`sub_type` for schema name required.")
@@ -124,7 +195,7 @@ def determine_schema(entity_type, sub_type):
     return schema
 
 
-def _get_response(metadata, entity_type, sub_type, validate_uuids, pathname=None):
+def _get_response(metadata, entity_type, sub_type, validate_uuids, pathname=None) -> dict:
     if validate_uuids == '1':
         response = validate_records_uuids(metadata, entity_type, sub_type, pathname)
     else:
@@ -137,40 +208,67 @@ def _get_response(metadata, entity_type, sub_type, validate_uuids, pathname=None
     return response
 
 
-def get_col_id_name_by_entity_type(entity_type):
+def get_col_id_name_by_entity_type(entity_type: str) -> str:
+    """
+    Returns the tsv id column name for the given entity type
+
+    Parameters
+    ----------
+    entity_type str entity type
+
+    Returns st name of the column in the tsv
+    """
     if equals(entity_type, 'Sample'):
         return 'sample_id'
     else:
         return 'source_id'
 
 
-def get_sub_type_name_by_entity_type(entity_type):
+def get_sub_type_name_by_entity_type(entity_type: str) -> str:
     if equals(entity_type, 'Sample'):
         return 'sample_category'
     else:
         return 'source_type'
 
 
-def supported_metadata_sub_types(entity_type):
+def supported_metadata_sub_types(entity_type: str) -> list:
     if equals(entity_type, 'Source'):
         return ['Human']
     else:
         return ["Block", "Section", "Suspension"]
 
 
-def validate_records_uuids(records, entity_type, sub_type, pathname):
+def validate_records_uuids(records: list, entity_type: str, sub_type, pathname: str):
+    """
+    Validates the uuids / SenNet ids of given records.
+    This is used for bulk upload so that ancestor ids referenced by the user in TSVs
+    are found to actually exist, are supported and confirm to entity constraints.
+
+    Parameters
+    ----------
+    records list of records to validate
+    entity_type stt entity type
+    sub_type str sub-type of the entity
+    pathname str pathname of the tsv; always returned in the response for tracking and other re-validation
+
+    Returns Rest response containing results of validation
+    -------
+
+    """
     errors = []
     passing = []
     ok = True
     idx = 1
     for r in records:
+        # First get the id column name, in order to get SenNet id in the record
         id_col = get_col_id_name_by_entity_type(entity_type)
         entity_id = r.get(id_col)
-        url = commons_file_helper.ensureTrailingSlashURL(current_app.config['ENTITY_WEBSERVICE_URL']) + 'entities/' + entity_id
+        entity_url: str = commons_file_helper.ensureTrailingSlashURL(current_app.config['ENTITY_WEBSERVICE_URL'])
         auth_helper_instance: AuthHelper = AuthHelper.instance()
         token = auth_helper_instance.getAuthorizationTokens(request.headers)
-        resp = requests.get(url, headers={f'Authorization: Bearer {token}',
-                                          'X-Application: ingest-api'})
+        resp = requests.get(f'{entity_url}entities/{entity_id}',
+                            headers={f'Authorization: Bearer {token}',
+                                     'X-Application: ingest-api'})
         if resp.status_code < 300:
             entity = resp.json()
             if sub_type is not None:
@@ -183,6 +281,7 @@ def validate_records_uuids(records, entity_type, sub_type, pathname):
                                                        f"on check of given `{entity_id}`. "
                                                        f"Supported `{'`, `'.join(supported_metadata_sub_types(entity_type))}`.",
                                                        idx, sub_type_col), dict_only=True))
+                # Check that the stored entity _sub_type matches what is expected (the type being bulk uploaded)
                 elif not equals(sub_type, _sub_type):
                     ok = False
                     errors.append(rest_response(
@@ -205,7 +304,8 @@ def validate_records_uuids(records, entity_type, sub_type, pathname):
         idx += 1
 
     if ok is True:
-        return rest_ok({'data': passing, 'pathname': pathname}, dict_only=True)
+        return rest_ok({'data': passing, 'pathname': pathname},
+                       dict_only=True)
     else:
         return rest_response(StatusCodes.UNACCEPTABLE,
                              'There are invalid `uuids` and/or unmatched entity sub types', errors,
@@ -238,16 +338,21 @@ def validate_metadata_upload():
         response = error
 
         if error is None:
+            if check_cedar(entity_type, sub_type, upload) is False:
+                id = get_cedar_schema_ids().get(sub_type)
+                return rest_response(StatusCodes.UNACCEPTABLE, 'Unacceptable Metadata',
+                                     f"Mismatch of \"{entity_type} {sub_type}\" and \"metadata_schema_id\". Valid id for \"{sub_type}\": {id}. "
+                                     f"For more details, check out the docs: https://docs.sennetconsortium.org/libraries/ingest-validation-tools/schemas")
+            path: str = upload.get('fullpath')
             schema = determine_schema(entity_type, sub_type)
-            path = upload.get('fullpath')
             validation_results = validate_tsv(path=path, schema=schema)
             if len(validation_results) > 2:
                 response = rest_response(StatusCodes.UNACCEPTABLE, 'Unacceptable Metadata',
                                          json.loads(validation_results), True)
             else:
                 records = get_metadata(upload.get('fullpath'))
-                response = \
-                    _get_response(records, entity_type, sub_type, validate_uuids, pathname=upload.get('pathname'))
+                response = _get_response(records, entity_type, sub_type, validate_uuids,
+                                         pathname=upload.get('pathname'))
                 if tsv_row is not None:
                     os.remove(upload.get('fullpath'))
